+++ src/docparser.cpp
@@ -708,7 +708,7 @@ int DocParser::handleAHref(DocNodeVariant *parent,DocNodeList &children,
                            const HtmlAttribList &tagHtmlAttribs)
 {
   AUTO_TRACE();
-  uint32_t index=0;
+  size_t index=0;
   int retval = RetVal_OK;
   for (const auto &opt : tagHtmlAttribs)
   {
@@ -778,7 +778,7 @@ void DocParser::handleLinkedWord(DocNodeVariant *parent,DocNodeList &children,bo
 
   const Definition *compound=0;
   const MemberDef  *member=0;
-  uint32_t len = context.token->name.length();
+  size_t len = context.token->name.length();
   ClassDef *cd=0;
   bool ambig;
   FileDef *fd = findFileDef(Doxygen::inputNameLinkedMap,context.fileName,ambig);
@@ -1570,7 +1570,7 @@ void DocParser::handleImg(DocNodeVariant *parent, DocNodeList &children,const Ht
 {
   AUTO_TRACE();
   bool found=FALSE;
-  uint32_t index=0;
+  size_t index=0;
   for (const auto &opt : tagHtmlAttribs)
   {
     AUTO_TRACE_ADD("option name={} value='{}'",opt.name,opt.value);
@@ -1663,9 +1663,9 @@ void DocParser::readTextFileByName(const QCString &file,QCString &text)
 
 //---------------------------------------------------------------------------
 
-static QCString extractCopyDocId(const char *data, uint32_t &j, size_t len)
+static QCString extractCopyDocId(const char *data, size_t &j, size_t len)
 {
-  uint32_t s=j;
+  size_t s=j;
   int round=0;
   bool insideDQuote=FALSE;
   bool insideSQuote=FALSE;
@@ -1713,7 +1713,7 @@ static QCString extractCopyDocId(const char *data, uint32_t &j, size_t len)
   {
     j+=9;
   }
-  uint32_t e=j;
+  size_t e=j;
   if (j>0 && data[j-1]=='.') { e--; } // do not include punctuation added by Definition::_setBriefDescription()
   QCString id(data+s,e-s);
   //printf("extractCopyDocId='%s' input='%s'\n",qPrint(id),&data[s]);
@@ -1728,9 +1728,9 @@ static QCString extractCopyDocId(const char *data, uint32_t &j, size_t len)
    do if ((i+sizeof(str)<len) && qstrncmp(data+i+1,str,sizeof(str)-1)==0) \
    { j=i+sizeof(str); action; } while(0)
 
-static uint32_t isCopyBriefOrDetailsCmd(const char *data, uint32_t i,size_t len,bool &brief)
+static size_t isCopyBriefOrDetailsCmd(const char *data, size_t i,size_t len,bool &brief)
 {
-  uint32_t j=0;
+  size_t j=0;
   if (i==0 || (data[i-1]!='@' && data[i-1]!='\\')) // not an escaped command
   {
     CHECK_FOR_COMMAND("copybrief",brief=TRUE);    // @copybrief or \copybrief
@@ -1739,9 +1739,9 @@ static uint32_t isCopyBriefOrDetailsCmd(const char *data, uint32_t i,size_t len,
   return j;
 }
 
-static uint32_t isVerbatimSection(const char *data,uint32_t i,size_t len,QCString &endMarker)
+static size_t isVerbatimSection(const char *data,size_t i,size_t len,QCString &endMarker)
 {
-  uint32_t j=0;
+  size_t j=0;
   if (i==0 || (data[i-1]!='@' && data[i-1]!='\\')) // not an escaped command
   {
     CHECK_FOR_COMMAND("dot",endMarker="enddot");
@@ -1763,7 +1763,7 @@ static uint32_t isVerbatimSection(const char *data,uint32_t i,size_t len,QCStrin
   return j;
 }
 
-static uint32_t skipToEndMarker(const char *data,uint32_t i,size_t len,const QCString &endMarker)
+static size_t skipToEndMarker(const char *data,size_t i,size_t len,const QCString &endMarker)
 {
   while (i<len)
   {
@@ -1778,7 +1778,7 @@ static uint32_t skipToEndMarker(const char *data,uint32_t i,size_t len,const QCS
     i++;
   }
   // oops no endmarker found...
-  return i<len ? i+1 : static_cast<uint32_t>(len);
+  return i<len ? i+1 : len;
 }
 
 
@@ -1786,7 +1786,7 @@ QCString DocParser::processCopyDoc(const char *data,size_t &len)
 {
   AUTO_TRACE("data={} len={}",Trace::trunc(data),len);
   GrowBuf buf;
-  uint32_t i=0;
+  size_t i=0;
   int lineNr = tokenizer.getLineNr();
   while (i<len)
   {
@@ -1794,7 +1794,7 @@ QCString DocParser::processCopyDoc(const char *data,size_t &len)
     if (c=='@' || c=='\\') // look for a command
     {
       bool isBrief=TRUE;
-      uint32_t j=isCopyBriefOrDetailsCmd(data,i,len,isBrief);
+      size_t j=isCopyBriefOrDetailsCmd(data,i,len,isBrief);
       if (j>0)
       {
         // skip whitespace
@@ -1858,10 +1858,10 @@ QCString DocParser::processCopyDoc(const char *data,size_t &len)
       else
       {
         QCString endMarker;
-        uint32_t k = isVerbatimSection(data,i,len,endMarker);
+        size_t k = isVerbatimSection(data,i,len,endMarker);
         if (k>0)
         {
-          uint32_t orgPos = i;
+          size_t orgPos = i;
           i=skipToEndMarker(data,k,len,endMarker);
           buf.addStr(data+orgPos,i-orgPos);
           // TODO: adjust lineNr
@@ -1880,7 +1880,7 @@ QCString DocParser::processCopyDoc(const char *data,size_t &len)
       lineNr += (c=='\n') ? 1 : 0;
     }
   }
-  len = static_cast<uint32_t>(buf.getPos());
+  len = buf.getPos();
   buf.addChar(0);
   AUTO_TRACE_EXIT("result={}",Trace::trunc(buf.get()));
   return buf.get();
